[0.4.6] keyboard input works
Canvas API 1.2.1
Window API 2.0
View API 2.0

+++ src/core/Window.cpp
@@ -111,7 +111,7 @@ fun Window::onMouseButton(i32 button, i32 action, i32 mods){
 }
 fun Window::onMouseScroll(f64 dx,f64 dy){
     if(mainView != null){
-        mainView->onScroll(dx,dy);
+        mainView->onScroll((f32)dx,(f32)dy);
     }
 }
